requirements.txt
fastapi>=0.111
uvicorn[standard]>=0.29
httpx>=0.27
pyyaml>=6.0
aiofiles>=23.0    # required for FileResponse
